personal_ai_coach/requirements.txt
streamlit>=1.40.0
openai>=1.59.0
python-dotenv>=1.0.0
